食事情報関連APIインターフェース設計書作成

+++ 基本設計/App/データベース設計書/ddl.sql
@@ -22,7 +22,7 @@ create table diet.t_user (
   , register_by CHAR(36)
   , register_date DATE
   , meal_type INT
-  , calorie FLOAT
+  , calorie INT
   , comment VARCHAR(100)
   , created_by CHAR(36) not null
   , created_at DATETIME not null
@@ -52,7 +52,7 @@ create table diet.t_meal (
   , register_by CHAR(36) not null
   , register_date DATE
   , meal_type INT
-  , calorie FLOAT
+  , calorie INT
   , comment VARCHAR(100)
   , created_by CHAR(36) not null
   , created_at DATETIME not null
